# internal/http/openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VBPClaw Gateway API",
    "description": "PostgreSQL multi-tenant AI agent gateway with WebSocket RPC + HTTP API.\n\n## Authentication\n\nAll endpoints require a Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <gateway-token-or-api-key>\n```\n\nYou can use either the **gateway token** (grants admin access) or an **API key** created via the API Keys endpoints (grants scoped access).\n\nIf no token is configured on the server, authentication is disabled (backward compatibility).\n\n## Common Headers\n\n| Header | Description |\n|--------|-------------|\n| `X-VBPClaw-User-Id` | External user ID for multi-tenant context |\n| `X-VBPClaw-Agent-Id` | Target agent ID (alternative to model prefix) |\n| `X-VBPClaw-Tenant-Id` | Tenant scope — UUID or slug (gateway token / cross-tenant API keys) |\n| `Accept-Language` | Locale for error messages (`en`, `vi`, `zh`) |\n\n## WebSocket Protocol\n\nConnect via `POST /ws` (upgrade). Protocol v3 uses frame types: `req`, `res`, `event`.\nFirst request must be `connect` with `{\"token\": \"...\", \"user_id\": \"...\", \"locale\": \"en\"}`.",
    "version": "0.2.0",
    "contact": {
      "name": "VBPClaw",
      "url": "https://github.com/kadiesnguyen/vbpclaw"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/chat/completions": {
      "post": {
        "tags": ["Chat"],
        "summary": "OpenAI-compatible chat completions",
        "description": "Send messages to an agent and get a response. Compatible with OpenAI's chat completions API.\n\nUse the `model` field to target a specific agent: `vbpclaw:<agent-id>` or `agent:<agent-id>`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["messages"],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Agent targeting: `vbpclaw:<id>` or `agent:<id>`. Defaults to `default`.",
                    "example": "vbpclaw:default"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["role", "content"],
                      "properties": {
                        "role": { "type": "string", "enum": ["system", "user", "assistant"] },
                        "content": { "type": "string" }
                      }
                    }
                  },
                  "stream": { "type": "boolean", "default": false },
                  "user": { "type": "string", "description": "External user ID" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Chat completion response (or SSE stream if stream=true)" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
...
    "/v1/tools/builtin/{name}": {
      "get": {
        "tags": ["Built-in Tools"],
        "summary": "Get built-in tool",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Tool details" } }
      },
      "put": {
        "tags": ["Built-in Tools"],
        "summary": "Update tool settings",
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Tool updated" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gateway token or API key"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or invalid token",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}